Q.48/mlops_pipeline.py
# MLOps Pipeline for Financial Services
# This script defines the core MLOps pipeline using MLflow, Airflow, and other tools
# for training, versioning, interpretability, deployment, and monitoring.
# Configurations for CI/CD, Docker, Kubernetes, and monitoring are included below.

import mlflow
import mlflow.sklearn
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, f1_score
import pandas as pd
import numpy as np
import shap
from fairlearn.metrics import MetricFrame, demographic_parity_difference
from fairlearn.reductions import ExponentiatedGradient, DemographicParity
import joblib
import os
from datetime import datetime
import logging
import boto3
from evidently import ColumnMapping
from evidently.report import Report
from evidently.metric_preset import DataDriftPreset, TargetDriftPreset
import prometheus_client as prom
import yaml

# Set up logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')

# ------------------- Dataset and Feature Versioning -------------------
# Description: Use DVC for dataset versioning and MLflow for experiment tracking.
# PII is anonymized using a custom anonymization function before storage.
# Datasets are stored in S3 with versioning enabled.

def anonymize_pii(df):
    """Anonymize PII columns (e.g., name, SSN) using hashing."""
    pii_columns = ['name', 'ssn']  # Example PII columns
    for col in pii_columns:
        if col in df.columns:
            df[col] = df[col].apply(lambda x: hash(str(x)) % 1000000)
    return df

def load_and_version_data(s3_bucket, s3_key, local_path='data.csv'):
    """Load data from S3, anonymize PII, and version with DVC."""
    s3 = boto3.client('s3')
    s3.download_file(s3_bucket, s3_key, local_path)
    df = pd.read_csv(local_path)
    df = anonymize_pii(df)
    
    # Version dataset with DVC
    os.system(f'dvc add {local_path}')
    os.system('dvc push')
    return df

# ------------------- Model Training and Validation -------------------
# Description: Train a RandomForest model, log metrics/parameters with MLflow,
# and perform hyperparameter tuning using Optuna.

def train_model(df, hyperparams=None):
    """Train a RandomForest model and log to MLflow."""
    X = df.drop(columns=['target'])
    y = df['target']
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
    
    with mlflow.start_run():
        # Set default hyperparameters or use tuned ones
        params = hyperparams or {'n_estimators': 100, 'max_depth': 10}
        model = RandomForestClassifier(**params)
        model.fit(X_train, y_train)
        
        # Evaluate
        y_pred = model.predict(X_test)
        accuracy = accuracy_score(y_test, y_pred)
        f1 = f1_score(y_test, y_pred)
        
        # Log parameters and metrics
        mlflow.log_params(params)
        mlflow.log_metrics({'accuracy': accuracy, 'f1_score': f1})
        mlflow.sklearn.log_model(model, 'model')
        
        # Save model artifact
        model_path = f'models/model_{datetime.now().strftime("%Y%m%d_%H%M%S")}.pkl'
        joblib.dump(model, model_path)
        mlflow.log_artifact(model_path)
        
        return model, X_test, y_test, model_path

# ------------------- Interpretability and Fairness Checks -------------------
# Description: Use SHAP for interpretability and Fairlearn for fairness analysis.

def interpretability_check(model, X_test):
    """Generate SHAP explanations and log to MLflow."""
    explainer = shap.TreeExplainer(model)
    shap_values = explainer.shap_values(X_test)
    shap.summary_plot(shap_values, X_test, show=False)
    shap_plot_path = 'shap_summary.png'
    plt.savefig(shap_plot_path)
    mlflow.log_artifact(shap_plot_path)
    return shap_values

def fairness_check(model, X_test, y_test, sensitive_feature='gender'):
    """Perform fairness analysis using Fairlearn."""
    y_pred = model.predict(X_test)
    metric_frame = MetricFrame(
        metrics={'accuracy': accuracy_score},
        y_true=y_test,
        y_pred=y_pred,
        sensitive_features=X_test[sensitive_feature]
    )
    dpd = demographic_parity_difference(y_test, y_pred, sensitive_features=X_test[sensitive_feature])
    
    # Log fairness metrics
    mlflow.log_metric('demographic_parity_difference', dpd)
    mlflow.log_dict(metric_frame.by_group.to_dict(), 'fairness_metrics.json')
    
    # Mitigate bias if needed
    if dpd > 0.1:  # Threshold for intervention
        mitigator = ExponentiatedGradient(model, DemographicParity())
        mitigator.fit(X_test, y_test, sensitive_features=X_test[sensitive_feature])
        mlflow.sklearn.log_model(mitigator, 'mitigated_model')
    
    return dpd

# ------------------- Model Monitoring -------------------
# Description: Use EvidentlyAI for drift detection and Prometheus for latency monitoring.

def monitor_model(model, reference_data, current_data):
    """Monitor for data drift and concept drift using EvidentlyAI."""
    column_mapping = ColumnMapping(
        target='target',
        numerical_features=reference_data.drop(columns=['target']).columns,
        categorical_features=None
    )
    
    drift_report = Report(metrics=[DataDriftPreset(), TargetDriftPreset()])
    drift_report.run(reference_data=reference_data, current_data=current_data, column_mapping=column_mapping)
    drift_report.save_html('drift_report.html')
    mlflow.log_artifact('drift_report.html')
    
    # Check for drift
    drift_detected = any(m['drift_detected'] for m in drift_report.as_dict()['metrics'])
    if drift_detected:
        logging.warning("Drift detected! Triggering rollback.")
        rollback_model()
    
    return drift_detected

# Prometheus metrics
request_latency = prom.Histogram('model_request_latency_seconds', 'Model inference latency')

def log_latency(start_time):
    """Log inference latency to Prometheus."""
    latency = (datetime.now() - start_time).total_seconds()
    request_latency.observe(latency)

# ------------------- Rollback Mechanism -------------------
# Description: Rollback to the last best-performing model based on F1 score.

def rollback_model():
    """Rollback to the last best-performing model."""
    runs = mlflow.search_runs(order_by=['metrics.f1_score DESC'])
    if not runs.empty:
        best_run_id = runs.iloc[0]['run_id']
        best_model_path = mlflow.get_artifact_uri(run_id=best_run_id, artifact_path='model')
        logging.info(f"Rolling back to model from run {best_run_id}")
        # Deploy best model (handled in deployment section)
        deploy_model(best_model_path)

# ------------------- Model Deployment -------------------
# Description: Deploy model as a RESTful microservice using Docker and Kubernetes.

def deploy_model(model_path):
    """Deploy model using Kubernetes (manifests defined below)."""
    # Update Kubernetes deployment with new model path
    with open('k8s/deployment.yaml', 'r') as f:
        deployment = yaml.safe_load(f)
    deployment['spec']['template']['spec']['containers'][0]['env'][0]['value'] = model_path
    
    with open('k8s/deployment.yaml', 'w') as f:
        yaml.safe_dump(deployment, f)
    
    # Apply deployment
    os.system('kubectl apply -f k8s/deployment.yaml')
    logging.info(f"Deployed model from {model_path}")

# ------------------- Main Pipeline -------------------
# Description: Orchestrate the pipeline using Airflow.

def run_pipeline():
    """Main pipeline execution."""
    # Load and version data
    df = load_and_version_data('my-bucket', 'data/customer_data.csv')
    
    # Train model
    model, X_test, y_test, model_path = train_model(df)
    
    # Interpretability and fairness
    interpretability_check(model, X_test)
    fairness_check(model, X_test, y_test)
    
    # Monitor
    reference_data = df.sample(frac=0.5, random_state=42)
    current_data = df.sample(frac=0.5, random_state=43)
    monitor_model(model, reference_data, current_data)
    
    # Deploy
    deploy_model(model_path)

if __name__ == '__main__':
    run_pipeline()