Q.48/requirements.txt
# Requirements for MLOps Pipeline
mlflow==2.15.0
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.0
shap==0.46.0
fairlearn==0.10.0
joblib==1.4.2
boto3==1.34.0
evidently==0.4.0
prometheus-client==0.20.0
gunicorn==22.0.0
pyyaml==6.0.1
optuna==3.6.0